requirements.txt
# Core Dependencies for Render Deployment
python-telegram-bot==20.7
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
aiohttp==3.9.1
requests==2.31.0
alpha-vantage==2.3.1
pybit==5.6.0
flask==3.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
structlog==23.2.0
pandas==2.1.4
numpy==1.25.2
psycopg2-binary==2.9.10
pycryptodome==3.23.0
